refactor(orchestrator): relocate module-level helpers into orchestration/orchestrator-helpers.ts (seam 25, #1526)
Behavior-preserving move of all 62 module-level declarations (1,036 LOC
incl. docs): recall-mode planning, recall snapshot types/parsers,
abort/race helpers, artifact recall limits, replay source-time slicing,
day-summary date utilities, and QMD startup checks. orchestrator.ts
re-exports every previously-public name so existing importers
(coordinators, tests, root shims) keep working unchanged; class-internal
uses import back from the helpers module.

Also dedupes extraction-run.ts's private copies of throwIfRecallAborted
/raceRecallAbort in favor of the shared helpers (AGENTS.md rule 8), and
repoints the three source-grep tests that pinned moved helper bodies.

orchestrator.ts: 5,890 -> 4,864 LOC.
